requirements/local.txt
ipython==7.19.0
django-extensions==3.0.9
django-cprofile-middleware==1.0.5
django-debug-toolbar==3.2.1
Werkzeug==1.0.1
